use the config.ini file and lib.py source file for both of the web scraper programs
