+++ config.ini
@@ -1,0 +1,5 @@
+[popular_certs]
+certs = Comptia Network+,Comptia Security+,CCNA
+
+[IT_positions]
+positions = Network Engineer,Network Administrator,Cybersecurity Engineer
